+++ src/rrd/Makefile
@@ -8,6 +8,7 @@ SRC += src/rrd/stack.c
 SRC += src/rrd/pretty_collapse.c
 SRC += src/rrd/pretty_affix.c
 SRC += src/rrd/pretty_bottom.c
+SRC += src/rrd/pretty_nested.c
 SRC += src/rrd/pretty_redundant.c
 SRC += src/rrd/pretty_skippable.c
 SRC += src/rrd/pretty_roll.c
